fix: refactor logging to use the utils.logger file and ensure separation of concerns
Refs: #10
