+++ src/tests/conftest.py
@@ -1,14 +1,6 @@
 import pytest 
 from main import create_app
-from logging import Handler, LogRecord
-
-class LogCaptureHandler(Handler):
-    def __init__(self):
-        super().__init__()
-        self.records = []
-
-    def emit(self, record: LogRecord):
-        self.records.append(record)
+from utils.logger import LogCaptureHandler
 
 @pytest.fixture
 def app():
