+++ src/utils/logger.py
@@ -1,0 +1,9 @@
+from logging import Handler, LogRecord
+
+class LogCaptureHandler(Handler):
+    def __init__(self):
+        super().__init__()
+        self.records = []
+
+    def emit(self, record: LogRecord):
+        self.records.append(record)
